requirements.txt
cryptography==3.4.7
Flask==2.0.1
oto==1.0.1
PyMySQL==1.0.2
python-dotenv==0.17.1
SQLAlchemy==1.4.18
